requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bcrypt==4.3.0
bidict==0.23.1
certifi==2025.7.14
click==8.2.1
dnspython==2.7.0
docutils==0.22
email-validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
ifaddr==0.2.0
itsdangerous==2.2.0
jinja2==3.1.6
markdown-it-py==4.0.0
markdown2==2.5.4
markupsafe==3.0.2
mdurl==0.1.2
multidict==6.6.3
nicegui==2.22.1
orjson==3.11.1
passlib==1.7.4
propcache==0.3.2
pscript==0.7.7
pydantic==2.11.7
pydantic-core==2.33.2
pygments==2.19.2
pyjwt==2.10.1
python-dotenv==1.1.1
python-engineio==4.12.2
python-multipart==0.0.20
python-socketio==5.13.0
pyyaml==6.0.2
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
sentry-sdk==2.34.1
shellingham==1.5.4
simple-websocket==1.1.0
sniffio==1.3.1
sqlalchemy==2.0.42
sqlmodel==0.0.24
starlette==0.47.2
typer==0.16.0
typing-extensions==4.14.1
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
vbuild==0.8.2
watchfiles==1.1.0
websockets==15.0.1
wsproto==1.2.0
yarl==1.20.1
